Remove excess logging
