+++ server/session/session-storage.ts
@@ -34,14 +34,12 @@ export default class SessionStorage {
   }
 
   private getStore(app: Application) {
-    logger.info('redis enabled', config.get('features.redis.enabled'));
     if (config.get('features.redis.enabled')) {
       logger.info('Using Redis session store', config.get('secrets.opal.redis-connection-string'));
       const client = createClient({
         url: config.get('secrets.opal.redis-connection-string'),
       });
 
-      logger.info('Connecting to Redis');
       client.connect().catch(logger.error);
 
       client.on('error', (err) => {
