+++ server/sso/sso-login.ts
@@ -4,9 +4,9 @@ import config from 'config';
 import axios from 'axios';
 
 const INTERNAL_USER_LOGIN = `${config.get('opal-api.url')}/internal-user/login-or-refresh`;
+const logger = Logger.getLogger('login');
 
 export default async (req: Request, res: Response, next: NextFunction) => {
-  const logger = Logger.getLogger('login');
   const env = process.env['NODE_ENV'] || 'development';
   const hostname = env === 'development' ? config.get('frontend-hostname.dev') : config.get('frontend-hostname.prod');
   const url = `${INTERNAL_USER_LOGIN}?redirect_uri=${hostname}/sso/login-callback`;
