+++ server/sso/sso-logout-callback.ts
@@ -2,9 +2,9 @@ import { NextFunction, Request, Response } from 'express';
 import { Logger } from '@hmcts/nodejs-logging';
 import config from 'config';
 
-export default (req: Request, res: Response, next: NextFunction) => {
-  const logger = Logger.getLogger('logout');
+const logger = Logger.getLogger('logout');
 
+export default (req: Request, res: Response, next: NextFunction) => {
   req.session.destroy((err) => {
     if (err) {
       logger.error('Error destroying session', err);
